Refactor to make stun resolve abstract
