+++ internal/stun/stun.go
@@ -1,0 +1,11 @@
+package stun
+
+import (
+	"github.com/google/wire"
+	"github.com/tjjh89017/stunmesh-go/internal/ctrl"
+)
+
+var DefaultSet = wire.NewSet(
+	NewResolver,
+	wire.Bind(new(ctrl.StunResolver), new(*Resolver)),
+)
